// src/apis/models/Volo_FileManagement_Files_FileDescriptorDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_FileManagement_Files_FileDescriptorDto = {
    readonly extraProperties?: Record<string, any> | null;
    id?: string;
    creationTime?: string;
    creatorId?: string | null;
    lastModificationTime?: string | null;
    lastModifierId?: string | null;
    directoryId?: string | null;
    name?: string | null;
    mimeType?: string | null;
    size?: number;
    concurrencyStamp?: string | null;
};